Fix unsafe cast from uint16_t* to size_t*.
This cast is not safe on 32-bit systems.

Once we drop support for AVR we can replace uint16_t with uint32_t
everywhere and life will be great.

+++ controller/lib/serdes/serdes.h
@@ -7,9 +7,9 @@ typedef void (*GuiAckHandler_t)(GuiAck &);
 
 bool serdes_encode_status_packet(ControllerStatus controller_status,
                                  uint8_t *tx_buffer, uint8_t tx_buffer_len,
-                                 size_t *encoded_len);
+                                 uint16_t *encoded_len);
 bool serdes_decode_incomming_packet(uint8_t *rx_buffer, uint8_t rx_buffer_len,
-                                    size_t encoded_len,
+                                    uint16_t encoded_len,
                                     GuiAckHandler_t gui_ack_handler,
                                     CommandHandler_t command_handler);
